[FIX/#2] 온보딩 패키지 재조정
