+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:theme="@style/Theme.ANDANDROID"
         tools:targetApi="31">
         <activity
-            android:name=".feature.login.LogInActivity"
+            android:name=".feature.onboarding.login.LogInActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -23,7 +23,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".feature.signup.SignUpActivity"
+            android:name=".feature.onboarding.signup.SignUpActivity"
             android:exported="false" />
         <activity
             android:name=".feature.main.MainActivity"
