+++ app/src/main/java/org/sopt/and/feature/main/home/navigation/HomeNavigation.kt
@@ -1,18 +1,18 @@
-package org.sopt.and.feature.profile.navigation
+package org.sopt.and.feature.main.home.navigation
 
 import androidx.navigation.NavController
 import androidx.navigation.NavOptions
 import kotlinx.serialization.Serializable
 import org.sopt.and.navigation.MainTabRoute
 
-fun NavController.navigateToProfile(
+fun NavController.navigateToHome(
     navOptions: NavOptions? = null,
 ) {
     navigate(
-        Profile,
+        Home,
         navOptions
     )
 }
 
 @Serializable
-data object Profile : MainTabRoute
+data object Home : MainTabRoute
