+++ app/src/main/java/org/sopt/and/feature/main/save/SaveScreen.kt
@@ -1,4 +1,4 @@
-package org.sopt.and.feature.home
+package org.sopt.and.feature.main.save
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.material3.Text
@@ -6,19 +6,19 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 
 @Composable
-fun HomeRoute(
+fun SaveRoute(
     modifier: Modifier = Modifier
 ) {
 
-    HomeScreen()
+    SaveScreen()
 }
 
 
 @Composable
-fun HomeScreen(
+fun SaveScreen(
     modifier: Modifier = Modifier,
 ) {
     Column {
-        Text("Home")
+        Text("Save")
     }
 }
