+++ app/src/main/java/org/sopt/and/feature/onboarding/login/LogInActivity.kt
@@ -1,4 +1,4 @@
-package org.sopt.and.feature.login
+package org.sopt.and.feature.onboarding.login
 
 import android.app.Activity
 import android.content.Intent
@@ -16,7 +16,7 @@ import dagger.hilt.android.AndroidEntryPoint
 import org.sopt.and.designsystem.component.appbar.OnboardingTopBar
 import org.sopt.and.designsystem.theme.ANDANDROIDTheme
 import org.sopt.and.feature.main.MainActivity
-import org.sopt.and.feature.signup.SignUpActivity
+import org.sopt.and.feature.onboarding.signup.SignUpActivity
 
 @AndroidEntryPoint
 class LogInActivity : ComponentActivity() {
