+++ app/src/main/java/org/sopt/and/feature/onboarding/login/LogInState.kt
@@ -1,4 +1,4 @@
-package org.sopt.and.feature.login
+package org.sopt.and.feature.onboarding.login
 
 data class LogInState(
     val id: String = "",
