+++ app/src/main/java/org/sopt/and/feature/onboarding/signup/SignUpSideEffect.kt
@@ -1,4 +1,4 @@
-package org.sopt.and.feature.signup
+package org.sopt.and.feature.onboarding.signup
 
 sealed class SignUpSideEffect {
     data class SignUpError(val isIdError: Boolean) : SignUpSideEffect()
